B - hash count must be uint64_t, not size_t, so that it doesn't change in width between 32- and 64-bit systems.

+++ include/hset_encoder.h
@@ -116,13 +116,13 @@ size_t length_hhnn(
 
 size_t write_hhnn_data(
   const RecordFieldDescriptor& hi,
-  size_t hash_count,
+  uint64_t hash_count,
   char* out
 );
 
 size_t write_hhnn(
   const RecordFieldDescriptor& hi,
-  size_t hash_count,
+  uint64_t hash_count,
   char* out
 );
